pyproject.toml
[tool.poetry]
name = "chatapath"
version = "0.1.0"
description = "A messaging system"
authors = ["Olena Gut <lena.gut1993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
pytest = "^8.2.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
ipython = "^8.24.0"
factory-boy = "^3.3.0"
faker = "^25.0.1"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
